ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
pypdf==4.3.1
python-pptx==0.6.23
python-docx>=0.8.11

# Advanced PDF extraction & OCR
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Storage / Supabase access (optional)
supabase>=2.5.0

# AI/ML models for generation and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Optional visual term extraction (CLIP)
git+https://github.com/openai/CLIP.git

# For testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
jsonschema>=4.21.1
